chatbot/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
autopep8==2.3.2
beautifulsoup4==4.14.3
black==25.12.0
build==1.3.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docopt==0.6.2
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.8.0
fastar==0.8.0
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
google-auth==2.45.0
google-genai==1.56.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
Js2Py==0.74
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-ollama==1.0.1
langchain-openai==1.1.6
langchain-qdrant==1.1.0
langchain-text-splitters==1.1.0
langsmith==0.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.31
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
numpy==2.2.6
ollama==0.6.1
openai==2.14.0
opencv-python==4.12.0.88
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pathspec==0.12.1
peewee==3.18.3
pillow==12.1.0
pip-tools==7.5.2
pipwin==0.5.2
platformdirs==4.5.1
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyjsparser==2.7.1
pyparsing==3.3.1
pypdf==6.5.0
PyPrind==2.11.3
pyproject_hooks==1.2.0
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytokens==0.3.0
pytube==15.0.0
pytz==2025.2
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
seaborn==0.13.2
sentry-sdk==2.48.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8.1
SpeechRecognition==3.14.5
SQLAlchemy==2.0.45
starlette==0.50.0
sympy==1.14.0
ta==0.11.0
TA-Lib==0.6.8
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typer==0.21.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
yfinance==1.0
yt-dlp==2025.12.8
zstandard==0.25.0
